+++ limit-orders.php
@@ -6,7 +6,7 @@
  * Author URI:  https://nexcess.net
  * Text Domain: limit-orders
  * Domain Path: /languages
- * Version:     1.1.2
+ * Version:     1.2.0
  *
  * WC requires at least: 3.9
  * WC tested up to:      4.0